Rename penaltymodel.lp to penaltymodel.generation
